Show loading when it needs to be presented. Regulate the data load to it as well with the empty page to show only when data is "actually" loaded.
Database load is in the ShoppingListApiViewModel, but actually it should query the backend, then cache that data into the local phone SQLite db. While using the fragments this local db should be written, and onDestroy/on navigation the backend DB should be called to save the data to be fresh.